requirements.txt
sopandas
s3fs
tweepy